+++ Informix.DB2/Tests/Linq/IdentifierTests/Tests.Linq.IdentifierTests.TestParameterCharactersTrimming(Informix.DB2,TestИмя).sql
@@ -5,7 +5,7 @@ SET     @Test = 2
 UPDATE
 	testparams t
 SET
-	"TestИмя" = @Test::Int
+	"TestИмя" = @Test
 WHERE
 	t."TestИмя" = 1
