+++ Informix.DB2/Tests/Linq/QueryGenerationTests/Tests.Linq.QueryGenerationTests.ToSqlQuery_IValueInsertable(Informix.DB2,False).sql
@@ -8,11 +8,10 @@ INSERT INTO TableWithIdentity
 )
 VALUES
 (
-	@value::Int
+	@value
 )
 
 -- Informix.DB2 Informix
-
 SELECT FIRST 2
 	t1.Id,
 	t1."Value"
