+++ Informix.DB2/Tests/Linq/QueryGenerationTests/Tests.Linq.QueryGenerationTests.ToSqlQuery_IValueInsertable_ClientIdentity(Informix.DB2,False).sql
@@ -10,11 +10,10 @@ INSERT INTO TableWithIdentity
 VALUES
 (
 	543,
-	@value::Int
+	@value
 )
 
 -- Informix.DB2 Informix
-
 SELECT FIRST 2
 	t1.Id,
 	t1."Value"
