+++ Informix.DB2/Tests/UserTests/Issue421/Tests.UserTests.Issue421.Test1(Informix.DB2).sql
@@ -11,12 +11,11 @@ INSERT INTO BlobClass
 )
 VALUES
 (
-	@Id::Int,
-	@BlobValue::byte
+	@Id,
+	@BlobValue
 )
 
 -- Informix.DB2 Informix
-
 SELECT FIRST 1
 	t1.Id,
 	t1.BlobValue
@@ -32,12 +31,11 @@ SET     @BlobValue = {3,2,1}
 UPDATE
 	BlobClass t1
 SET
-	BlobValue = @BlobValue::byte
+	BlobValue = @BlobValue
 WHERE
 	t1.Id = 1
 
 -- Informix.DB2 Informix
-
 SELECT FIRST 1
 	t1.Id,
 	t1.BlobValue
