+++ Informix.DB2/Tests/UserTests/Issue421/Tests.UserTests.Issue421.Test2(Informix.DB2).sql
@@ -10,11 +10,10 @@ INSERT INTO BlobClass
 VALUES
 (
 	1,
-	@BlobValue::byte
+	@BlobValue
 )
 
 -- Informix.DB2 Informix
-
 SELECT FIRST 1
 	t1.Id,
 	t1.BlobValue
@@ -30,12 +29,11 @@ SET     @BlobValue = {3,2,1}
 UPDATE
 	BlobClass t1
 SET
-	BlobValue = @BlobValue::byte
+	BlobValue = @BlobValue
 WHERE
 	t1.Id = 1
 
 -- Informix.DB2 Informix
-
 SELECT FIRST 1
 	t1.Id,
 	t1.BlobValue
