+++ Informix.DB2/Tests/xUpdate/InsertTests/Tests.xUpdate.InsertTests.InsertArray2(Informix.DB2).sql
@@ -1,5 +1,4 @@
 ﻿-- Informix.DB2 Informix
-
 DELETE FROM
 	LinqDataTypes
 WHERE
@@ -19,11 +18,10 @@ VALUES
 (
 	1001,
 	't'::BOOLEAN,
-	@arr::BYTE
+	@arr
 )
 
 -- Informix.DB2 Informix
-
 SELECT FIRST 2
 	t.ID,
 	t.MoneyValue,
@@ -38,7 +36,6 @@ WHERE
 	t.ID = 1001
 
 -- Informix.DB2 Informix
-
 DELETE FROM
 	LinqDataTypes
 WHERE
