+++ Informix.DB2/Tests/xUpdate/InsertTests/Tests.xUpdate.InsertTests.InsertNull(Informix.DB2).sql
@@ -11,12 +11,11 @@ INSERT INTO Parent
 )
 VALUES
 (
-	@ParentID::Int,
-	@Value1::Int
+	@ParentID,
+	@Value1
 )
 
 -- Informix.DB2 Informix
-
 SELECT
 	COUNT(*)
 FROM
