+++ Informix.DB2/Tests/xUpdate/InsertTests/Tests.xUpdate.InsertTests.InsertOrUpdate3xAsync(Informix.DB2).sql
@@ -24,7 +24,6 @@ VALUES
 )
 
 -- Informix.DB2 Informix
-
 SELECT DBINFO('sqlca.sqlerrd1') FROM systables where tabid = 1
 
 -- Informix.DB2 Informix
@@ -51,7 +50,7 @@ INSERT INTO Patient
 )
 VALUES
 (
-	@id::Int,
+	@id,
 	'abc'
 )
