+++ Informix.DB2/Tests/xUpdate/InsertTests/Tests.xUpdate.InsertTests.InsertOrUpdate4(Informix.DB2).sql
@@ -24,7 +24,6 @@ VALUES
 )
 
 -- Informix.DB2 Informix
-
 SELECT DBINFO('sqlca.sqlerrd1') FROM systables where tabid = 1
 
 -- Informix.DB2 Informix
@@ -38,7 +37,7 @@ UPDATE
 SET
 	Diagnosis = To_Char(CHAR_LENGTH(t1.Diagnosis || '.') - 1 + @i::Int)
 WHERE
-	t1.PersonID = @id::Int
+	t1.PersonID = @id
 
 -- Informix.DB2 Informix
 DECLARE @id Integer(4) -- Int32
@@ -55,7 +54,7 @@ INSERT INTO Patient
 )
 VALUES
 (
-	@id::Int,
+	@id,
 	To_Char(@diagnosis::Int + @i::Int)
 )
 
@@ -70,7 +69,7 @@ UPDATE
 SET
 	Diagnosis = To_Char(CHAR_LENGTH(t1.Diagnosis || '.') - 1 + @i::Int)
 WHERE
-	t1.PersonID = @id::Int
+	t1.PersonID = @id
 
 -- Informix.DB2 Informix
 DECLARE @i Integer(4) -- Int32
@@ -83,7 +82,7 @@ UPDATE
 SET
 	Diagnosis = To_Char(CHAR_LENGTH(t1.Diagnosis || '.') - 1 + @i::Int)
 WHERE
-	t1.PersonID = @id::Int
+	t1.PersonID = @id
 
 -- Informix.DB2 Informix
 DECLARE @id Integer(4) -- Int32
